+++ CHANGELOG.md
@@ -15,6 +15,7 @@ planned for 2025-07-01
 
 - [config] Allow to change module order for final renderer (or dynamicaly with CSS): Feature `order` in config. (#3762)
 - [clock] Added option 'disableNextEvent' to hide next sun event (#3769)
+- [clock] Implement short syntax for clock week (#3775)
 
 ### Changed